refat-tacometro
corrige erro grave onde servidor emitia dado computado de rpm.
define mensagem Notificada como um pseudo json nao padrao menor.
segundo campo de atributo c marca contagem de rotacoes foi adicionado.

+++ servidor/Tacometro.cpp
@@ -34,7 +34,7 @@ int medirTempoEntreDoisPulsos()
   Servidor.ino deve impedir execução desta função
   para duracaoRotacao < 10 evitando retorno 0.1
 */
-float obterRpm( int duracaoRotacao )
+int obterRpm( int duracaoRotacao )
 {
   if ( duracaoRotacao > 10 )
   {
